I have gotten the size of the low level and high level buffers to reflect their values and be in sync when they should be, and not when they shouldn't. I can write a full AbstractedCanPacket to a CAN Frame, and ostensibly it works exactly as it should.
This means that creating a full CAN Frame from the high level Abstracted Can Packet is working.
We can set the nodeID, the Priority, and add MiniPackets until the buffer is full, and write
that to the CAN Frame. From there it is trivial to send the built CAN Frame.

However, constructing the AbstractedCanPacket back from a received CAN Frame is still
unfinished.