I have created the CanBitBuffer class, which is basically a subset of the code from AbstractedCanPacket which deals only with reading/writing to the CAN Packet.
It works, ostensibly, and I have actually fixed two bugs. One is that the CanPrint method
was printing incorrectly when the buffer was exactly full.

The other was that writing to the CAN packet was overflowing past msg.buf[7] if msg.ext == 0.
The issue was that maxBufferSize is assuming that msg.ext == 1.